moved functions from BaseAnalysis to a separate file
